Support JWT authentication
Add a new option `--auth-type` to switch between "basic" and "jwt"
authentication methods.

In the case of jwt, the `--auth-config` should point to a valid
JWKS (either a path to a file, or an URL). The JWKS will be reloaded
at regular interval.

The JWT token will be extracted either from the header
`Authorization: Bearer <TOKEN>` or from the `Token: <TOKEN>`.
It should contain a claim called `namespaces` (list of strings)
and be signed with the key in the JWKS that matches the
`kid` found in the JWT header.

The defaults are set so that there is no breaking changes for
existing users.

More information is available in the readme.