Optionally support AWS signature
This is useful is the prometheus being proxied is an APS -
Amazon Prometheus Service. In this case, every request need to
be signed, and the signature depends on the request URL.
As the query is modified on the fly, this needs to be handled by
the proxy itself.

See the readme for more information.

Note that the proxy must preserve path from the original request, as
APS URL is of the form https://aps-workspaces.<region>.amazonaws.com/workspaces/<UUID>/api/v1

Some tests are added for `reverse.go` (incomplete).